Fix parsing errors by using correct access values
The problem of missing `children` members was solved by setting to an
empty list and in a future commit I will add a recursive async function
that goes through and fills any missing gaps. I believe by separating
the sync and async parts of the parsing we can keep the code less complex
and more readable.

We are currently relying on the JsonMap type using dynamic rather than
Object? and will change that when we update json_types but for now it
makes for more readable code and the analyzer will take us to each place
we need to update when the time comes.