I implemented missing member for Page and added tests. The class is now "final" as part of a test to see whether making classes "final" works for the goals of the package (I think I will provide mock classes) eg. being able to add 'extension' packages that provide further funcitonality for different contexts, eg. outputting a string description of a page in a format suitable to post to Discord.
I finished the `getAllBlockChildren` which repeatedly calls
`getBlockChildren` until there are no more children and returns
everything in a single list.

I updated the example to use the new getAllBlockChildren method, because
it's a bit cleaner so makes a better example I think.

I updated Parent and the relevant subtypes to use a cleaner pattern,
with no "type" members and just a single "id" member in all subtypes.
The plan is to make Parent a "sealed" class to provide an ADT but for
now that messed up existing tests so I've left pulling the trigger for a
future PR.